+++ R/autoplot.R
@@ -48,6 +48,9 @@
 autoplot.bench_mark <- function(object,
   type = c("beeswarm", "jitter", "ridge", "boxplot", "violin"),...) {
 
+  # Just convert everything to a character first
+  object$expression <- as.character(object$expression)
+
   res <- tidyr::unnest(object)
   p <- ggplot2::ggplot(res)