+++ NEWS.md
@@ -1,5 +1,7 @@
 # bench 1.0.1.9000
 
+* `scale_bench_time()` and `scale_bench_bytes()` now allow you to use a non-logarithmic scale.
+
 # bench 1.0.1
 
 * Add support for macOS versions prior to 10.12